requirements.txt
Django==5.0.9
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
celery[redis]==5.3.4
django-celery-beat==2.5.0
whitenoise==6.6.0
Pillow==10.1.0
qrcode[pil]==7.4.2
segno==1.6.1
reportlab==4.0.7
drf-spectacular==0.26.5
PyJWT==2.8.0
cryptography==41.0.7
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
mypy==1.7.1
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
ruff==0.1.6
black==23.11.0
pre-commit==3.5.0
